# rewrite bucket, use a struct instead of map